rename of action file to exerciseActions.js and changed all imports for that
